redid on-map nav buttons, #abstract behaviour
the up button is gone, it was confusing to users. its functionality will be handled by the nav bar, that is to go on top of the map.

the expand button has a label now, for better visibility. the label updates dynamically.

the button uses a forked leaflet.button.js now.

the abstract should now always retreat when a gebiet is selected.